Allow Boost 1.38.0 instead of 1.40.0 at minimum. Use Boost by default on Windows.

+++ include/kNet/Types.h
@@ -25,12 +25,12 @@ typedef unsigned long dword;
 
 typedef unsigned char u8; ///< a single byte: 0-255.
 typedef unsigned short u16; ///< 2 bytes: 0 - 65535.
-typedef unsigned long u32; ///< 4 bytes: 0 - 4,294,967,295 ~ 4000 million or 4e9.
+typedef unsigned int u32; ///< 4 bytes: 0 - 4,294,967,295 ~ 4000 million or 4e9.
 typedef unsigned long long u64; ///< 8 bytes: 18,446,744,073,709,551,615 ~1.8e19.
 
 typedef signed char s8; ///< a single byte: -128 - 127.
 typedef signed short s16; ///< 2 bytes: -32768 - 32767.
-typedef signed long s32; ///< 4 bytes signed: max 2,147,483,647 ~ 2000 million or 2e9.
+typedef signed int s32; ///< 4 bytes signed: max 2,147,483,647 ~ 2000 million or 2e9.
 typedef signed long long s64; ///< 8 bytes signed. 9,223,372,036,854,775,807 ~ 9e18.
 
 typedef unsigned long ulong;
